feat(evaluation): identify telemetry readings by reading id with double-precision pm

+++ src/main/java/com/claircore/evaluation/infrastructure/persistence/jpa/embeddables/ParticulateMatterPersistenceEmbeddable.java
@@ -4,5 +4,5 @@ import jakarta.persistence.Embeddable;
 
 /** Storage shape of {@code ParticulateMatter}; component names are addressed by the entity. */
 @Embeddable
-public record ParticulateMatterPersistenceEmbeddable(Integer pm1_0, Integer pm2_5, Integer pm10) {
+public record ParticulateMatterPersistenceEmbeddable(Double pm1_0, Double pm2_5, Double pm10) {
 }
